+++ src/lib.rs
@@ -30,6 +30,8 @@
 //!   value, so a missing field error is generated. `Option<T: FromMeta>` uses this to make options optional without requiring
 //!   `#[darling(default)]` declarations, and structs and enums can use this themselves with `#[darling(from_none = ...)]`.
 //!   This takes either a path or a closure whose signature matches `FromMeta::from_none`.
+//! * **Generate `syn::parse::Parse` impl**: When deriving `FromMeta`, add `#[darling(derive_syn_parse)]` to also generate an impl
+//!   of the `Parse` trait.
 //!
 //! ## Forwarded Fields
 //! All derivable traits except `FromMeta` support forwarding some fields from the input AST to the derived struct.